Card Indices
It is now possible to access the index of a card from within the card's template. This is accessible via the $index property of the card card reference like {{card.$index}}. This should close issue #6.

Note that this index refers to the index of the original collection and not to the (generated) column index. The existing tests are extended to cover this feature.