Add possibility to set card template with attribute string or with inner HTML instead of using an external template file.
The card templates can now be provided by using a cardTemplateString attribute (which uses the attribute value as the template) or by omitting any template attributes (which uses the inner HTML of the directive as the template). Please see the extended README for details of the syntax and functionality. Specs for these new features are included. This should fix both aspects of issue #7.